fix, remove class ServerSideAuthorization
